Adding parallel retrievers, improved classification model
